Clarify the the example in the README is not literal
It is reasonable to expect that a user may take an example of a shell
command from a README, copy it, paste it into their shell, and then
run it.  I know users do this because *I* have done it who knows how
many times.  It is in no way a good, safe practice, nor one that I
recommend or support.  But my life is the source of endless examples of
me engaging in behavior which I explicitly condemn.