Fixes two typos in the LICENSE document
These typos appear because of a program I use as part of an over-arching
'create a new programming project' piece of software.  I was going to
write a small patch to fix the bug that caused one of the typos, so I
looked at the original author of the code which caused the problem:

The original author was me, from October of 2015.

Simply out-fucking-standing.

Special-thanks: Eric James Michael Ritz circa 2015
Special-thanks: El Psy Congroo